requirements.txt
# Web framework
fastapi==0.115.0
uvicorn==0.30.0
python-multipart==0.0.9       # File uploads (audio, images, video)

# Database
sqlalchemy==2.0.35
alembic==1.13.2

# Data validation & settings
pydantic==2.9.0
pydantic-settings==2.5.0

# AI / LLM
openai==1.51.0                 # Transcription (gpt-4o-mini-transcribe)
anthropic==0.37.0              # Intent classification + vision (Claude)

# Google APIs
google-api-python-client==2.149.0   # Calendar, Tasks, Gmail
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1         # OAuth flow for Google

# SMS
twilio==9.3.0                  # Text messages to Johnny

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.27.0

# Media processing
Pillow==10.4.0                 # Image handling
